Make skip_internal_error work by design, not by accident
This function depended on such an extremely clever encoding of SC+OFFSETs
that I don't think there was any awareness of just how tied to encoding
it was, because it's so insane that it could not have been designed.
The comment asking whether it could be right to treat the number of args
as a number of bytes was right only if your arg SC was descriptor-reg,
and only for x86[-64].

With a few exceptions, the only time we need to skip an internal error
is with BOGUS-ARG-TO-VALUES-LIST-ERROR. That error takes one argument, and
the argument's SC is DESCRIPTOR-REG which has SCN 0.
SCN 0 is encodable in the low 2 bits of an SC+OFFSET without using any
extension bits (bits 23 and up), leaving 5 bits availabe for the register
without overflowing the largest value encodable as exactly one byte of a
LEB128-encoded varint. So, yeah, that's crazy, because as soon as I tried
making a continuable trap with an UNSIGNED-REG arg, the skipping went wrong
and caused a crash after CERROR returned control.

So this change correctly decodes LEB128. But note the new comment saying
that ARM64 is now right, but only by accident and only in the current usage.

Incidentally, the other continuable error trap is on Alpha, HPPA, MIPS,
and RISCV in their implementations of LIST/LENGTH which allows ignoring a
dotted last cons.  I'll bet we don't have a test of that. But I tried it
on MIPS, and it works. However, all of those architectures avoid calling
skip_internal_error(), and just skip some fixed amount of bytes.