x86oid: optimize CASE with symbol keys
For CASE forms having >= 6 symbols, try to find a perfect hash from the bits
of SYMBOL-HASH and rewrite the CASE over small integer keys. If we can't find
a good hash, or collisions cause the expansion to be too complex, then fall
back to the conventional expansion.

We can usually find a perfect hash with a single LDB. A notable exception
is with symbols whose print names are STRING=.
And we can easily arrange for all symbols referenced in compiled code to have
a precomputed hash. This was already done a while ago, so only the comment is
revised to mention the new relevance.

This optimization only makes sense if CASE over integers is better than a
string of EQ comparisons, so don't do any of this for non-x86 architectures
as they do not currently support jump tables.