Simplify %assemble-sections
The use of packages as a namespace for opcodes never panned out the
way I wanted, so this change somewhat does away with that, and especially
the abhorrent use of of lower-case symbols to avoid conflict with the CL
package. (And it never worked for more than one assembler to be loaded)
Opcodes can be in any package; with care, it's possible to compare them
by EQ when writing assembly passes, as long as the package is consistent.