Transform (MEMBER x '(symbol ...)) to CASE
Whether or not it is used as a predicate, the code is generally better
than an if/else chain, and it's actually far denser than the inline
code when the list becomes huge. As long as the symbols are perfectly
hashed, the code size is insensitive to the number of symbols.