Fix error & warning in make-target-2-load
Cold-error defines a stub for WARN which does nothing by default though
selectable by *!cold-warn-action* which seems always to be NIL under the
presumption that we're unable to handle warnings in that stage.
The real WARN isn't installed until after loading fasls.
I believe this is on account of referring to the type named CLASS
which isn't defined until PCL is built.