Slightly improve the code around a fixme
It's not clear that this fixes anything, but it's not worse, and it's a
small step toward a change that I want to make to avoid zero-filling while
pseudo-atomic. SIMPLE-VECTOR is the only object type that presents a
technical challenge due to the need to avoid scavenging random bytes.
Other objects can be zeroed at leisure, if at all. COPY-BIGNUM, for example,
takes all the bytes from the source bignum right away.