Initialize lisp TLS slots without help from C
except for the ones that C absolutely has to know about.

This fixes the problem that create_thread_struct() was not GC-safe
if any assignment referred to a symbol other than T or NIL.

Initialization forms for thread-local specials are no longer
restricted to be (OR FIXNUM SYMBOL).

Also address some inefficiencies:
- the TLS index slot in SYMBOLS lacked a type specifier.
- nowhere did we declare that *CURRENT-THREAD* is of type THREAD.

This change really made me question why REINIT after a failure from
SAVE-LISP-AND-DIE creates a new main thread struct while leaving
other things alone, but I tried not to alter pre-existing behavior.

There is a new slightly annoying problem with 'slam.sh' though,
depending on what it needs to recompile.