Make unknown-type instances quasi-interned
For any atomic specifier that is unknown, always return the same internal object,
using a weak hash-table to implement that behavior.

This is mainly an aid to examining behavior of the cross-compiler, though possibly
also user-compiled code, though nobody would know that. (Even though we allow
deferral of "undefined" warnings related to types, it's probably almost always
a bad idea to refer to types that aren't known.)

Anyway, with this patch you can clearly see that INTERPRETED-FUNCTION persists as
unknown through make-host-2, but then goes away by the time make-target-2 is done.
Afaict, it is made worse by (though not caused directly by) the :UNPARSE method
on an intersction type, which is to say, attempting to print an intersection type
moves the problem around, because its unparser parses COMPILED-FUNCTION which is
(AND FUNCTION (NOT INTERPRETED-FUNCTION)).