Improve error message about misuse of * as type specifier
Also, reject (VALUES *) as per CLHS. This is a behavior change
but it did not affect any regression tests.

Afaict, we're too liberal about * still. It is nothing more than syntax to
omit a positional arg, and as such, the FUNCTION type compound parser is
probably still wrong except when the entire arg list is *.
Within the type machinery, *WILD-TYPE* has its legitimate uses of course.