Fix sb-show, broken due to removal of NEW-OBJECT-ID
sb-show is for debugging cold-init, not for seeing how compiler IR
structures are related.  Other ways exist of doing the latter,
and it was arbitrary that of the ~100 structure types used in the
compiler, only about 12 of them got a unique ID.

Also fix a problem in compiling src/code/aprof.