Banish CROSS-FLOAT-INFINITY-KLUDGE.
* Teach the math journal about multiple values.
* Add more hard-coded cases to cross-float.
* Allow irrational function usage in the host.
* Remove #+sb-xc-host paths for float type derivers.
  - Doesn't seem to change the size of the cold core, but I suspect
  there are cases where our system code adds manual type declarations in
  lieu of missing type derivers.
  - Incidentally, the old CMUCL #-propagate-fun-type and
  #-propagate-float-type feature paths corresponding to the
  #+sb-xc-host paths mentioned in the WHN comment have been gone in
  CMUCL since as early as 2002.

Now that hard coded special cases and the math journal are used much
more, with float trap masking being important in one case, it appears
that for the sake of maintainance, having in addition to/instead a
tiny soft float library modeling our target float operations (outside
of irrational functions) would be sensible, removing the need for hard
coded cases involving -0.0, infinities and much of the journal. Such a
soft float library could have uses outside of cross compilation as
well, for example in helping emulate traps on platforms without
hardware support for traps or infinities.