mips: use a chosen register, not $gp, for jump relaxation
It is extremely weird that J and JAL would secretly clobber $gp.
The $at (assembler temp) register is more suited to that.
Otoh, if our use of $gp as a scratch register (not in accordance with the
ABI) was actually ok, then we should have declared it in mips/vm.lisp as a
register generally usable by lisp, and not just for this one weird trick.