Simplify ppc SB-VM::ALLOCATION function
For non-threaded builds on all architetures:
* Write an 8-word unboxed array at the beginning of static space.
  Move NIL up by that amount.

For PPC with gencgc and no threads:
* In lieu of using the C gc_alloc_region structure, use the unboxed array
  at the start of static space to contain such a structure.
  Teach GC about that alloc_region.

For all architectures:
* Correct MAP-ALLOCATED-OBJECTS so that it does not visit the array
  that overlays a 'struct alloc_region'

This change remove the largest impediment to implementing :SB-DYNAMIC-CORE
universally which is that reading a C gc_alloc_region structure from Lisp
requires another indirection due to not wiring in any C data addresses.
The generic allocator does not always have a spare register in which
to perform the extra indirection.