Allow relocating fixedobj space when starting ELFinated image
The only thing preventing relocation was that Lisp deliberately never stored
fixups from varyobj space into fixedobj space, because in normal usage those
spaces are displaced by a predetermined amount, namely the size of the lower
subspace. But an ELF core has to know how to adjust instructions if fixedobj
space is not remapped where expected.  e.g. in a CALL from immobile code to
a generic-function with self-contained trampoline, which is a fixedobj.
Editcore now rewrites %CODE-FIXUPS with additional fixups as needed.