Add -no-as-needed for ppc linker
The default for some toolchains is to leave out of the executable the
information necessary to load 'libm.so' (generally 'foo.so'), absent any
apparent link-time reference to it. But we need libm despite no use of it
from our C runtime, so opt out of the pruning step.

It's funny to put "-Wl," in OS_LIBS, but it is an order-sensitive flag
(counteracted by the opposite flag if seen later), and our GNUmakefile
adds it just before "-lm" which is exactly the right place, as it
pertains to the following library or libraries on the linker invocation.