Default vop-name from :TRANSLATE when unspecified
A mandatory name had the unsatisfying aspect of often not being particularly
useful for anything, but differing in some stupid way such as being spelled
with an extra character, just adding to confusion instead of reducing it.
This change isn't any particularly more risky than requiring a name in terms
of unintentionally clobbering (via redefinition) an existing vop of the same,
since we have no meaningful way to detect redefinition anyway.