Remove a stray #+sb-doc
Loading introspect from source not via asdf-stub would get a warning
about a removed feature.

It remains to be decided how to strip docstrings from contribs if desired.
The automated technique in make-host-2 doesn't suffice. A reasonable thing
might be for LOAD to remove docstrings in %DEFUN etc when loading a system
package if the base image was built with #-sb-doc.