Define LAYOUT differently in host and target
The host layout discards some slots that are used only for implementing
various aspects of the runtime such as TYPEP optimizations.

A corollary of this is that the GC bitmap can be hoisted into trailing raw
slots of the layout, which would be problematic in the host, as the
implementation will rely on having raw slots and variable-length instances.