ppc64: optionally use SIGILL for traps
This is basically a kludge. In a multi-thread test there are a million places
where you'd have to manually type "signal SIGTRAP" in gdb to proceed into
traps that you want to deliver to the application.
Combined with the fact that the gcc farm machines disallow (via ptrace_scope)
attaching gdb to a stuck process, and emulators are too slow to expose issues
related to hardware thread scheduling, there is very litle chance of making
progress on ppc64 without some alternative to sigtrap.
The previous hack (now removed) was not only unsafe, but didn't deal with
do_pending_interrupt traps, or error traps for that matter.