Use sb-xc:typep in FIND-CONSTANT and make it less sucky
Storing any INSTANCE other than PATHNAME in an EQUAL hash-table is not useful.
This change doesn't fix anything per se other than that inefficiency.

The real problems are documented in the newly added notes above the code.
I could not parse the comment concerning "retaining identity under EQUAL"
so I removed that. I think it meant the opposite: numbers do NOT have to
retain identity. It was a completely illogical claim.