Fix SB-PCL::PROBE-CACHE
It was broken so badly (by my 1b9728423c which attempted to optimize
out a CONS) that any key failing to hit its first line would miss.
(How could that cause no failures?)

In addition to fixing the bug, change the empty slot marker from a
symbol to the unbound marker; and don't use (NOT SYMBOLP) to test for
key validity- use %INSTANCEP which is a quicker and better test.
Remove some tagbody/go style coding as well.