Use layout-info, not bitmap, in #n= reader
Most uses of the bitmap should be avoided from lisp in anticipation of
two changes: one to use a trailing word to store a stable hash value,
and one to place the bitmap into trailing raw words of the layout.

Also fix a latent bug with funcallable-instances. The raw words holding
the embedded trampoline code are indistinguishable from fixnums, but that
need not be the case. It just worked out that way by sheer luck.