Default to using :BLOCK-COMPILE :SPECIFIED for self build.
This allows us to use START-BLOCK and END-BLOCK declarations freely
within our own code. There is also a slight codegen change here, where
global functions with self calls are now eagerly resolved and local
call themselves, instead of going through an FDEFN. This may be a
slight performance win.

:SPECIFIED does make incremental redefinition marginally less dynamic
though, in that recompiling FOO on the fly while FOO is calling itself
will not cause the the running FOO to call the newly redefined
FOO. This is actually not observable when loading the system to
develop on, however, since you need to load everything anyway, and
redefining FOO will instantly give you back the dynamic redefinition
behavior.