handle EQ conses in CDR chains when dumping
Before dumping the CAR of a given cons, check to see whether the cons
itself has been seen previously in this file compilation.  We cannot
rely on the EQ table alone, because we might have non-trivial shared
structure between two lists (whose CDR chain is dumped as a single
list rather than N conses), so have a new table to store associations
between conses and the list object that has been dumped.

(The code to do this is strongly reminiscent of the code to handle
circularity within a single object, which requires backpatching in a
way that structure sharing doesn't.  It might be possible to share
logic or entries between the circularity-table and this new
cons-table.)