Unpessimize SXHASH on LAYOUT
Lazy stable hashing caused a regression.

Additionally, remove 2 layout flag bits. I thought we needed to pick off CTYPE,
but that was due to badness in the compiler's coalescing where it tried to find
structure instances in the "similar" table.
And no point wasting 1 bit for layout-of-layout, it's better as an EQ test.