Fix "cast from pointer to integer of different size" warning
The way I saw this was that a non-unicode SBCL got a character decoding error
in RUN-PROGRAM on the C compiler whose output contained
  stack-alignment-offset.c: In function ‘stack_alignment_offset’:
where the quote mark is (CODE-CHAR #x2018) = #\LEFT_SINGLE_QUOTATION_MARK.

The SBCL error was only visible in the parallel test executor which seems to
run with LANG=en_US.UTF-8 instead of LANG=C. I don't know why the discrepancy.
But, hey, it was a good reason to get a locale setting wrong!