Fix fd leak with closefrom(2)-like interface.
The current approach has two problems: first, it can leak an fd if a high
fd is opened in the parent, and the max-fd limit is lowered before the
fds are closed prior to exec. The second problem is that the code has to
issue a close call for every possibly-valid fd value, which can introduce
significant delays on systems where the max-fd limit is high and close
calls are slow (this has been observed in the wild in at least one cloud-
based environment).

To remedy these problems, we use a closefrom(2)-like interface to close all
open fds before execing the child process. On BSD systems that have a
real closefrom(2) system call, we use that; for other platforms, if the
F_CLOSEM fcntl is available, we use that; otherwise, we attempt to walk
the /dev/fd/ and /proc/self/fd/ directories. Since these may not be
available (e.g. in a chroot), if all of these fail, we use a fallback
implementation.

The fallback implementation is nearly the same as the previous code,
except that it uses the F_MAXFD fcntl to determine that maximum fd if
it's available. This will avoid leaking fds in the child, although it
won't address the poor performance of the fallback method.

To use this interface, the pipe fd left open in the child process for
signalling purposes is duped to the lowest available fd (3), and the
closefrom wrapper is then used to close all open fds above that. Setting
FD_CLOEXEC on the pipe fd now occurs after the pipe fd is moved, since
dup2 does not preserve fd flags.

committer: Stas Boukarev <stassats@gmail.com>