Avoid multiple definition error with GCC 10 on Windows.
GCC 10 now compiles with -fno-common by default, which triggers a linker
error if a variable is defined in a header and shared in more than one
object file to be linked. DEAD_MUTEX is shared, and triggers the error;
declare it with extern and move the variable into the pthreads_win32.c
file to avoid it.

committer: Stas Boukarev <stassats@gmail.com>