Bypass ALLOCATE-INSTANCE in obsolete instance trap
Calling it does not seem to accord with the AMOP where it says that new slot
storage is created, and does not say new "instance". (I don't have the exact
wording, and my paper copy is currently inaccessible).

None of {ABCL, CLISP, ECL} call ALLOCATE-INSTANCE in the obsolete instance
protocol. Other implementations do, so it's seemingly neither wrong nor
required to do so; but it's definitely unportable for user code to depend
on one way or the other.