Teach elfcore.test.sh about non-linux OSes and reenable.
This adds a new -sbcl-libs flag to tests/run-compiler.sh for adding
flags necessary to link the runtime and uses it in elfcore.test.sh.
Ideally the makefiles from src/runtime/ would be used to avoid
duplication, but this is a simpler and less invasive change.

Additionally, don't try to execute an sbcl from /tmp on OpenBSD, RWX
mmap() restrictions mean that anywhere outside the source directory is
unlikely to work.