More of less of WITH-LOCKED-SYSTEM-TABLE
This makes a few weak table mutexes nonrecursive which was extra problematic
because of the combined outer WITH-LOCKED-... coupled with automatic locking of
synchronized tables plus the fact that we essentially ignore the :SYNCHRONIZED
option of MAKE-HASH-TABLE when :WEAKNESS is specified, meaning that you can't
create an *unsynchronized* weak table. Implicit locking was the behavior from
git rev f318d0b165, but that was done as an artifact of the implementation
and not a contractual obligation.

But as we know, users tend to expect aspects of an API that are observable to
stay the same, at least until we loudly tell them otherwise. We don't know if
magical threadsafety of weak tables is something users have come to rely on.
So even if I could make weak tables multi-reader concurrent without locking
(with a lot of effort), we would need to announce that henceforth weak tables
behave as any other table that lacks the :synchronized arg or explicitly gives
it as NIL. That's an OK thing to do, but more than this change demands.
So I employed a hack to workaround the ignoring of the :SYNCHRONIZED option
rather than trying to respect it and fix all problems entailed by such fix.