Remove the hash-vector from EQL hash-tables
It's not used in any of the table operations, aside from rehash.
(No speed improvement comes from calling #'= to decide to call #'EQL)

Having no hash-vector decreases memory consumption by 15%.
The savings should outweigh the slight penalty in recalculating hashes
due to GC, as well as in GC to detect which keys are address-sensitive.