Make FINDHASH-WEAK immune to infinite looping
When an address-sensitive key is not found, at most one rehash needs to
be done, but a spectacularly unlucky scheduling of GC could have caused it
to repeatedly rehash. This bug was present prior to the function as it exists
in its current form- all variations of GETHASH could loop, because any
concurrent reader could mess up all chains for all other readers. In the new
algorithm, a rehash-without-growing is always associated with a linear search
for the key that triggered rehash. It's no more work to do both at once.
For a weak table, this algorithm takes advantage of the fact that we disallow
multi-readers (for now anyway) which makes the logic that much simpler.

This change also cleans up some leftover mess which is that element 1 of the
pairs vector was either known as the "needs-rehash" flag or the "epoch". It is
now consistently named the "rehash-stamp" which is orthogonal to the bug fix.