Reinstate lowtag check elision in some type vops
The elision logic was fine, but I wanted the resulting assembly code to be
as robust as it could be in the face of an "object" that does not have any
lisp type, namely the unbound marker. So now it is mostly harmless to try
to load a widetag from the unbound marker, because it is almost a heap object
but not- it's sort of like NIL but with a whacky lowtag.