Remove *ALL-THREADS-LOCK*
It's not needed to protect the AVL tree.

It was overloaded to not only do that but also prevent creation of new threads
which is better done with the make-thread lock; as well as prevent thread exit,
for which its use was completely horrible.

Since each thread already has a mutex for interruptions, use that for
SYMBOL-VALUE-IN-THREAD (and INTERRUPT-THREAD) to ensure that the memory
exists (in the former case) or the pthread ID exists (in the latter).
There is absolutely no reason to look in the C variable all_threads,
unless you're using one of the platforms/configs that I don't understand.