Prevent JOIN-THREAD on foreign threads
When the thread returns from lisp back to C, we say that it has finished
which may or may not be true. And the result is NIL no matter what.
So not only can't you get the value, there's no way to ensure that the
pthread completed using our API in which THREAD-JOIN does not directly
entail a pthread_join() on the underlying thread.