Remove use of TRUENAME (via PROBE-FILE) from OPEN
This could possibly cause behavior change if options involving renaming are
supplied for a file that had symlinks in the path to it. But in plain usage,
opening any path other than exactly as requested violates the principle of
least surprise. Hence if the new and old behaviors differ, the new is correct.