win32: make pipes larger
When SB-EXT:RUN-PROGRAM is asked to return streams for stdin and/or
stdout on Windows, it creates pipes with a very small 256-byte buffer.
The smaller the buffer, the easier it becomes to enter deadlock,
particularly because FORCE-OUTPUT blocks when flushing the stream
buffer onto the pipe.

An example of how things might go wrong:

(let* ((process (sb-ext:run-program "program" nil
                                    :input :stream
                                    :output :stream
                                    :wait nil))
       (stdin (sb-ext:process-input process)))
  ;;
  ;; program writes out a startup banner larger than 256 bytes and blocks
  ;;
  (write-line "<a command larger than 256 bytes>" stdin)
  (force-output stdin)) ; blocks writing to the pipe, leading to deadlock

On the one hand, this might be considered a feature, an early warning
of sorts, that forces users to set up a more robust scheme such as
reading and writing from different threads. On the other hand, why
should Windows be the sole beneficiary of such a feature?

This patch changes SB-WIN32::WINDOWS-PIPE to use the system's default
pipe size, usually 4 KB.

committer: Stas Boukarev <stassats@gmail.com>