Add some timing and other diagnostics to deadlock.impure
After removal of the session lock on thread startup (git rev. faa46781)
and the all-threads lock (in lisp, not C - git rev. 1dbad9cb) it tests
nothing like what it originally meant to.
But it is nonetheless is interesting code because with suitable modifications
(mainly increasing the number of threads) it can be made to hit "interrupt
nesting depth exceeded" and/or mmap failure.