Deal with weird shells
/bin/sh on some GCC farm machines is a symlink to 'dash', and 'dash' has a
strange (wrong?) behavior of starting a process that has a child that you
don't know about. It does this when using "exec" and "<<" together.
Then the input provider (the one writing the text of the here-document)
is a child of SBCL. This may not be POSIX-compliant, but what do I know.