Implement fcb-threads test for win32
This change cranks up the number of GCs which causes it to flake on macOS.
All indications are that there's an OS bug - if it is observed that the GC
signal is pending in detach_os_thread(), the call to sigwait() sometimes
does not clear the signal. Usually it works. I can't discern a pattern.