win32: remove more thread-related garbage
Don't model the run state or implement join or detach. SB-THREAD:JOIN-THREAD
is its own thing using a Lisp semaphore, not pthread_join(), and OS threads
are freed automatically when the last handle is closed, so that's easy.

And this fixes a memory leak. There was no free() to match the calloc()
of 'struct pthread_thread', The only free() call pertained to foreign threads
there were "noticed" and then "unnoticed".