Re-allow concurrent JOIN-THREAD of 1 target thread
Discussion w/xof concluded that while this might not be particurly good to allow
in the API, it's easy enough to, and the consequence of disallowing it is bad:
we don't know whether users do this, and we can't present a reasonable error
at runtime (:concurrent-join-thead?) since the manifestation was to hang.

So even if it was purely an accident of implementation that it worked, it is
better to preserve that; however, the docstring suggests not doing something
which your C library can't, and gives us leeway for altering the behavior.

Add another missing (unrelated) test manifest entry as well.