win32: abolish struct pthread_thread
The pthread wrapper seemed to be a solution looking for a problem.
Just track the signal mask (useless as it is) anywhere in one of the
all-too-many places that we stash thread-specific bits.

Change inline asm to gcc intrinsics while we're at it.

Tested on Windows 10 Enterprise Eval for 64-bit. Untested on 32-bit.