Change interrupt-thread to use SIGURG
This change intertwines some issues:
 * which signal should be/not-be used for nefarious reasons
 * was ignoring a non-thread-targeted SIGPIPE accidental?

To the first point: The Go lang people have thought about it and reached
a conclusion that SIGURG is the best choice of signal to steal for a similar
purpose as ours (https://golang.org/src/runtime/signal_unix.go)

To the second: it is important to allow users to write lisp code that follows
the principle of least surprise when compared to C code without asking that
they infer the myriad ways in which SBCL goes contrary to Unix convention.

While this represents a possible departure from behavior introduced in rev
d4478142 (14 years ago), as with so many changes of the time, there was zero
motivation nor a test case, and it only later came to overlap with usurpation
of SIGPIPE for INTERRUPT-THREAD, which incidentally disregards an empty queue.

So this restores a sound approach in which we give control, not policy,
and also happens to correct a bug from last month's release - a handler
will never satisfy fixnump() because it gets passed to C as a raw word.