win32: add :WINDOW argument to SB-EXT:RUN-PROGRAM
Allows win32 users to decide whether the subprocess window should be
displayed and how it should be displayed.

Defaults to not affect how the subprocess window is displayed which
matches the previous behaviour. In the default behaviour, console
subprocesses will display a command line window when the SBCL runtime
is a GUI application.

committer: Stas Boukarev <stassats@gmail.com>