Finish removing dubious safepoint logic (follow-up to 1cb0a8e3)
The sb-sprof trick had two subtle flaws because a safepoint in a nested context
does not imply that it is safe to GC. It would be wrong for GC to close the
normal per-thread boxed region if it was in the midst of doing an allocation.
And that's if you can figure out how to avoid self-deadlock on free_pages_lock
without resorting back to using pseudo-atomic.