Remove fill-pointer-p primitive object slot
Use a header bit instead. The bit can't be on for a simple-array,
but it's always ok to read it.

Not only does this reduce conditional branching, it yields some space
savings for string-output-stream buffers at their default initial size.
Formerly such strings took 10 words of non-simple header plus 2 words of
simple header plus 8 data words. The 10 word header contained a padding
word, so removing 1 word gives a net decrease of 2 words,
and 2 words out of 20 is a 10% decrease.