Improve WIDETAGS-FROM-UNION-TYPE
After we "flatten" a type test that involves unions of unions, it generated
worse code than would have been generated had we tested each of the builtin
types separately. Unfortunately builtins get "lost" in types such as
(OR ARRAY COMPLEX SIMD-PACK) but we can do a little better than to repeat
the lowtag test N times over. %OTHER-POINTER-SUBTYPEP-P just needs a full list
of the widetags that it should look for.