Fix ASAN spurious failure - lp#1904779
Due to our workaround for https://sourceware.org/bugzilla/show_bug.cgi?id=1780
there were sigcontexts containing uninitialized bytes, and the pthread_sigmask
sanitizing interceptor does not know that the OS did not write all bytes of the
sigcontext which it places on your stack to begin with.

It does no good to add NO_SANITIZE all over our code because pthread_sigmask
still crashes, and I could see no way to call the non-sanitized entry point
(which would be the right solution) with the sanitizer generally enabled.

So I hacked at this to discover a minimal set of loci to forcibly unpoison
memory and yield 20,000 passing runs. But the solution still gives one the
uneasy feeling that I threw random darts at the problem.
And this still fails locally, but not in the cloud, which I attribute
to a different compiler options. That's enough hammering on it for now.