Begin removing CL function redefs from simple-streams
Define the CL: functions in such a way that they can handle simple-streams
if the contrib module is loaded on demand. This eliminates some helper
functions such as ANSI-STREAM-WRITE-STRING whose sole reason for being
was that simple-streams wanted to call it, which it no longer does.

This cleanup is being done not because it matters whether simple-streams
are usable; on the contrary it needs to be done so that we don't have to
define bits of stream.lisp assuming that it may be stomped on later
by a badly behaved contrib module. Whether the contrib works is, meh.