Use trivial early methods for some ansi stream operations
The ones which are the subject of this change generally had to perform
all of the following definitions to create one final definition:
 > (declaim (inline ansi-stream-op))
 > (defun ansi-stream-op (s) ...)
 > (defun stream-op (x) (if (ansi-stream-p x) (ansi-stream-op x) ...))
 > (fmakunbound 'stream-op)
 > (defgeneric stream-op (x))
 > (defmethod stream-op ((stream ansi-stream)) (ansi-stream-op stream))

All of that is replaced with just the defmethod and a small amount of
spporting code elsewhere. But despite having to wire in the names of these
into early-defmethod, it's still far simpler to do it this way.