Add more synchronization
Remove the futex-based finalizer thread trigger. At best it would sweep
some problems under a rug.

The "failed AVER: (THREAD-P THREAD)" was due to a bug in that change.
(The wait function in C returns no value)

I was able to reproduce hanging in 'saveN.test.sh' prior to this change,
and not able to after it, so things seem to be getting better.