Allow unbound slots in MAKE-LOAD-FORM-SAVING-SLOTS on a structure
If given :SLOT-NAMES which are a subset of slots, then place unbound markers in
un-dumped non-numeric slots. This improves consistency with the assumption that
slots always respect their type declaration, since unbound is no type; and
it is more rigorous than using a default of 0 (which is what the allocator does
"just because") and so is likely to help catch random usage errors.

For now, make-load-form-saving-slots is the only means by which unbound-marker
would appear in a structure slot. There has been some discussion of not using
NIL as the assumed default more generally (in the absence of a stated default
in the DEFSTRUCT form), but that would constitute a separate question.

A sampling of other Lisp implementations found the following behaviors
with respect to make-load-form-saving-slots for slots which had a value
at compile-time:
  ABCL: explicitly unassigned slots contain #<UNBOUND>
  ECL: ... contain SI:UNBOUND
  CCL: ... contain NIL
  CLISP: acts as though the structure constructor was called,
         and then overwrites any specified slots with dumped data

The behavior of CCL is dubious but understandable.
CLISP seems to have taken liberties with the spec.