Canonize unions of rational and integer types
Arrange for the canonical representation to have as wide as possible
an integer range, absorbing integers from the rational type if they
are adjacent.  If the rational type has open integral bounds, and
those integers are present in the integer type, replace the open
bounds by closed ones.

(Fixes a near-20-year-old FIXME)