Fix certainty in type-equality for unions
We test for type-equality in two unions by checking the SUBTYPEP
relationship in both directions.  If both are definitely SUBTYPEP each
other, then they are equal; if either is definitely not SUBTYPEP the
other, then they are definitely not equal.  However, if one is
SUBTYPEP the other but the other direction is not certain, we cannot
conclude anything with certainty.

Addresses some of lp#1916040.