Be less certain about type-equality with intersections.
Intersection types can be essentially any type in disguise, given the
presence of unknown or other sufficiently hairy types in the
intersection.  Investigate the individual intersection types, and if
any of them leads to uncertainty return NIL, NIL rather than
delegating to the other complex-= method (if any).

Annoyingly, although I saw this problem in the trace of a higher-level
subtypep operation, I can now not find arguments to subtypep that
tickle this.  The test included uses sb-kernel:type= explicitly.