Improve sb-sprof sampler robustness
* Treat all code as GC roots when sampling is in progress.

* Delay acquiring the lock until just before writing to the data buffer
  (after getting the backtrace)

* Don't collapse all "unknown" functions into one.
  Generate a unique placeholder for each distinct one.

* For x86-64, if the PC looks bad, don't save it as-is.
  Instead change it to the known "missing" value.

There is only one regression test (gc.impure) that shows "Unknown fn" in
its profile report after this. Fixing it doesn't seem too hard: *SPROF-DATA*
would need to be a static Lisp symbol so that GC can see if it is non-NIL,
and then treat it as if sb_prof_enabled were 1.
But I'm not convinced that the test is a valid use-case for the profiler.