Fix fun-end-breakpoint code blobs
They caused total destruction of your heap upon the next GC
because the jump-table prefix word read as nonzero.
Apparently not enough people use breakpoint-based tracing to have noticed,
and the test suite does not exercise it enough either.

Sparc and ppc64 seem not to fully work, but I think that's not new.