Disallow * in more type-parsing contexts
Most of the reason for allowing * had to do with the transforms for AREF etc
where an upgraded element type was blindly inserted into a THE form even if
it was unknown. No cast should be applied in that situation.

This partially fixes lp#1860919 but there is still an incorrect allowance
for * in (FUNCTION (*) T). The only valid use of * would be to indicate wholly
unknown arg types where the positional syntax requires something to be there.