Remove use of sbcl_pwd in contrib building.
Building the contribs doesn't need an absolute pathname for the source
tree; a relative pathname will do.

(Further, calling the environment variable SBCL_PWD is peculiarly
misleading in this context, since it's not the PWD of most processes
involved in make-target-contrib.sh; instead, it's a constant name for
the top-level source directory. In the hope of maximal clarity,
renamed the environment variable to SBCL_TOP, and SBCL-TOP in Lisp.)