Do not rely on fopcompiler to set FDEFNs ASAP in cold-init.
We can use FOP-FSET to always do this, since we were already using it
to bootstrap DEFUN linkage for block compiled DEFUNs, since block
compilation and fopcompilation do not mix. This allows us to remove
the fopcompiler linkage machinery dependent on %DEFUN and simplifies
logic in genesis and cold-init.

Things of note:
* Surprisingly, we don't need inline-expansions or argument dx
information before TLFs are run in cold-init. This might be because of
the current build order, or because it was never really necessary.
* COLD-FSET has been rewritten to be more like in CMU CL, since now
the linkage mechanism is basically exactly as it is in CMU CL. In
particular, there is a possibility of now COLD-FSETing with
closures. CMU CL seems to have code to support this, though it's not
clear to me whether it actually works or does anything useful.
* Remove the one usage of %DEFUN explicitly in the compiler. It's more
ugly to duplicate the definition but I think it's worth it to not rely
on %DEFUN being hardcoded to do cold-init magic.