Allocate minimal hash-tables even smaller
Some applications make zillions of hash-tables without ever adding a single
key to half of them. This change makes it more palatable to use hash-tables
in such cases - storage is allocated only when the first key is added.
This yields a 55% space savings for tables that remain empty for their life.