Make unqualified CHAR-CODE and CODE-CHAR refer to target values.
This way, it is much harder to accidentally leak host character codes
into the target image. The only instance where we use the host's
versions of these functions is for representing 0 in one place, which
is fine, since we don't want to create proxy objects for null
characters, and it's portable to use the host character in that
context anyway, since it will still get dumped as the null character.

This also incidentally fixes a real bug; we were using host character
codes for the +constituent-trait-table+ array when we definitely
should have been using target codes. That was probably borked on
non-ASCII hosts, not that any really exist anyway.