Fix subnormal integer-decode-float
INTEGER-DECODE-FLOAT was trying too hard. The reason is that it was
also used for DECODE-FLOAT and FLOAT-PRECISION, the former having a
constraint on the range of the significand, and the latter needing
to know how many bits are really significant.
But INTEGER-DECODE-FLOAT has no such constraint, and it is acceptable
to return the literal bits from the raw representation without trying
to rescale anything to eliminate leading zeros.
So it is now perfectly reliable. I can't say the same for DECODE-FLOAT.

This change *only* touches the 64-bit code, as 32-bit is (in my mind)
not worth bothering with.