Dump fops for package literals instead of using MAKE-LOAD-FORM.
Making the fasdumper dump literal package references directly makes
sense because we already have fops to do that. Plus, we no longer need
a MAKE-LOAD-FORM for it. Incidentally, both CLISP and CCL don't have
MAKE-LOAD-FORMs for PACKAGEs either, so they must also dump package
literals directly as well.

This makes both target and cross compiled fasls smaller.

Not only do we not need a special cross-compiler method to dump
package literals, we don't need to teach fop-funcall how to deal with
it anymore, though we do need to explicitly descriptorize the package
like we do for symbols when writing them out.

We also reduce dependence on the fopcompiler for making things work on
the cross-compiler and fasls smaller in the target compiler, which is
good for higher debug.