Don't do %DEFINE-INFO-TYPE at Genesis time.
!GLOBALDB-COLD-INIT can handle it fine, and this way we don't need to
duplicate an abstraction breaking cold version of %define-info-type
AND duplicate the action of registering the meta-info. The original
reason Genesis was made to handle it was because there was some other
deferral queue mechanism. We don't need to do that and can just keep
things simple with the ordinary DEFUN mechanism. Plus
!GLOBALDB-COLD-INIT can always be made to be done ASAP in cold-init,
so the bootstrap does not become any more brittle by doing so.