Rationalize the build order and remove some early- files.
* Many of the comments describing the dependencies in
build-order.lisp-expr were just plain obsolete, so deal with those.
* Try to move target-only files as late as possible, since it makes it
clear that the compiler doesn't depend on it in some weird way (and in
general, it shouldn't, because the compiler runs on the host).
* Try and move some "compiler/..." files into the compiler section of
the build order, to make it clear there's not some weird dependency
going on.
* Combine as many early-... files into their normal counterparts as
possible, to make clear that the build order dependency issues that
used to exist no longer do.
* Remove !PRED-COLD-INIT
* Move some miscellaneous stuff that clearly were in the wrong file to
a more sensible place.